+++ R/RcppExports.R
@@ -17,6 +17,10 @@ FusedLasso_cpp <- function(dataModel, intercept, regParam, controlFit) {
     .Call(`_quadrupen_FusedLasso_cpp`, dataModel, intercept, regParam, controlFit)
 }
 
+lava_dense_cpp <- function(dataModel, intercept, regParam, control) {
+    .Call(`_quadrupen_lava_dense_cpp`, dataModel, intercept, regParam, control)
+}
+
 ridge_cpp <- function(dataModel, intercept, regParam, controlFit) {
     .Call(`_quadrupen_ridge_cpp`, dataModel, intercept, regParam, controlFit)
 }