+++ R/Regularizers-R6Class.R
@@ -136,3 +136,44 @@ FusedLassoFit <- R6::R6Class(
     }
   )
 )
+
+#' Class "LavaFit"
+#' 
+#' Class of object returned by the fitting function [lava()]. Inherits fields
+#' and methods of [QuadrupenFit]
+#' 
+#' @seealso [QuadrupenFit], [lava()]
+#' 
+#' @export
+#' 
+LavaFit <- R6::R6Class(
+  classname = "LavaFit",
+  inherit = QuadrupenFit,
+  active  = list(
+    #' @field penalty character describing the regularizer/penalty
+    penalty = function(value) "lava",
+    #' @field sparse_coef sparse part of the  decomposition of the coefficients
+    sparse_coef = function(value) private$sparse_coef_,
+    #' @field dense_coef dense part of the  decomposition of the coefficients
+    dense_coef  = function(value) private$dense_coef_
+    ),
+  private = list(sparse_coef_ = NA, dense_coef_ = NA),
+  public  = list(
+    #' @description Initialize a [`LavaFit`] model
+    #' @param data a [`DataModel`] object
+    #' @param intercept a logical; should an intercept be included in the mode?
+    #' @param regParam a list with two elements, a vector and a scalar, for the regularization
+    initialize =  function(data, intercept, regParam) {
+      super$initialize(data, intercept, regParam)
+      private$optimizer <- lava_dense_cpp
+    },
+    #' @description function performing the optimization
+    #' @param control list controlling the optimization process    
+    fit = function(control) {
+      out <- super$fit(control)
+      private$sparse_coef_ <- out$beta
+      private$dense_coef_  <-  do.call(rbind, out$b)
+      private$beta <- private$sparse_coef_ + private$dense_coef_
+    }
+  )
+)